# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NEXIS Risk-IQ API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/{tenantId}/dashboard:
    get:
      operationId: getDashboard
      tags: [dashboard]
      summary: Get executive dashboard for a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /tenants/{tenantId}/kpis:
    get:
      operationId: getKpis
      tags: [dashboard]
      summary: Get KPIs for a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: KPI data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KpiSummary"
  /tenants/{tenantId}/departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments for a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a new department
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /tenants/{tenantId}/departments/{departmentId}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get department detail
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Department detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentDetail"
    put:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: departmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "200":
          description: Updated department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
...
  /tenants/{tenantId}/reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List available reports for a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a new report
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
  /tenants/{tenantId}/reports/{reportId}:
    get:
      operationId: getReport
      tags: [reports]
      summary: Get a specific report
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDetail"
  /tenants/{tenantId}/users:
    get:
      operationId: listUsers
      tags: [tenants]
      summary: List users for a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /tenants/{tenantId}/activity:
    get:
      operationId: listActivity
      tags: [dashboard]
      summary: Get recent activity feed for a tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEntry"
components:
  schemas:
...
    KpiSummary:
      type: object
      properties:
        totalRisks:
          type: integer
        controlsInProgress:
          type: integer
        overdueActions:
          type: integer
        highRiskItems:
          type: integer
        complianceRate:
          type: number
        appetiteBreaches:
          type: integer
        implementedControls:
          type: integer
        totalControls:
          type: integer
        awaitingReview:
          type: integer
        escalatedItems:
          type: integer
      required:
        - totalRisks
        - controlsInProgress
        - overdueActions
        - highRiskItems
        - complianceRate
        - appetiteBreaches
        - implementedControls
        - totalControls
        - awaitingReview
        - escalatedItems
    Dashboard:
      type: object
      properties:
        kpis:
          $ref: "#/components/schemas/KpiSummary"
        riskByDepartment:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentRiskSummary"
        implementationTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        riskLevelBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/RiskLevelCount"
        insights:
          type: array
          items:
            type: string
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityEntry"
      required:
        - kpis
        - riskByDepartment
        - implementationTrend
        - riskLevelBreakdown
        - insights
        - recentActivity
...
    RiskLevelCount:
      type: object
      properties:
        level:
          type: string
        count:
          type: integer
        percentage:
          type: number
      required: [level, count, percentage]
    Department:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        head:
          type: string
        totalRisks:
          type: integer
        totalControls:
          type: integer
        highRiskCount:
          type: integer
        overdueCount:
          type: integer
        complianceRate:
          type: number
        status:
          type: string
      required:
        - id
        - name
        - head
        - totalRisks
        - totalControls
        - highRiskCount
        - overdueCount
        - complianceRate
        - status
    DepartmentDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        head:
          type: string
        description:
          type: string
        complianceRate:
          type: number
        controls:
          type: array
          items:
            $ref: "#/components/schemas/Control"
        riskBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/RiskLevelCount"
        overdueItems:
          type: array
          items:
            $ref: "#/components/schemas/Control"
        accountability:
          type: array
          items:
            $ref: "#/components/schemas/AccountabilityEntry"
      required:
        - id
        - name
        - head
        - description
        - complianceRate
        - controls
        - riskBreakdown
        - overdueItems
        - accountability
...
    Report:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        description:
          type: string
        generatedAt:
          type: string
        period:
          type: string
        status:
          type: string
      required: [id, title, type, description, generatedAt, period, status]
    ReportDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        type:
          type: string
        description:
          type: string
        generatedAt:
          type: string
        period:
          type: string
        status:
          type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/ReportSection"
        summary:
          type: string
        recommendations:
          type: array
          items:
            type: string
      required:
        - id
        - title
        - type
        - description
        - generatedAt
        - period
        - status
        - sections
        - summary
        - recommendations
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        department:
          type: string
        avatar:
          type: ["string", "null"]
        lastActive:
          type: string
      required: [id, name, email, role, department, lastActive]
    ActivityEntry:
      type: object
      properties:
        id:
          type: string
        actor:
          type: string
        actorRole:
          type: string
        action:
          type: string
        item:
          type: string
        itemId:
          type: string
        timestamp:
          type: string
        result:
          type: string
        severity:
          type: string
      required:
        - id
        - actor
        - actorRole
        - action
        - item
        - itemId
        - timestamp
        - result
        - severity
...
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        head:
          type: string
        description:
          type: string
        actor:
          type: string
        actorRole:
          type: string
    ReportInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
        type:
          type: string
        description:
          type: string
        period:
          type: string
        actor:
          type: string
        actorRole:
          type: string